+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fragment, <></>を実装した。
 - <unwrap>を<fragment>に変更した
+- menu要素、param要素が欠けていたので追加
 
 ## [0.1.0-alpha.34] - 2025-10-22